feat(fe): change design
