+++ frontend/src/user/home/components/Carousel.vue
@@ -10,7 +10,7 @@ interface Slide {
   sub: string
   img: string
   imgAlt: string
-  color: 'green' | 'black' | 'white'
+  color: 'green' | 'black' | 'white' | 'yellow'
   href: string
 }
 const props = defineProps<{ slides: Slide[] }>()
@@ -18,12 +18,14 @@ const props = defineProps<{ slides: Slide[] }>()
 const bgColors: { [key: string]: string } = {
   green: 'bg-[#2e4e3f]',
   black: 'bg-[#333333]',
-  white: 'bg-[#ffffff]'
+  white: 'bg-[#ffffff]',
+  yellow: 'bg-[#f9de4a]'
 }
 const textColors: { [key: string]: string } = {
   green: 'text-white',
   black: 'text-white',
-  white: 'text-black'
+  white: 'text-black',
+  yellow: 'text-black'
 }
 
 const currentSlideIndex = ref(0)
@@ -46,51 +48,55 @@ const clickRight = () => {
 </script>
 
 <template>
-  <RouterLink :to="slides[currentSlideIndex].href">
-    <div
-      class="relative h-[450px] md:h-[400px]"
-      :class="bgColors[slides[currentSlideIndex].color]"
-    >
+  <div
+    class="relative mt-3 h-[400px] overflow-hidden rounded-3xl"
+    :class="bgColors[slides[currentSlideIndex].color]"
+  >
+    <div class="absolute bottom-0 left-0 right-0 z-50 mb-3 flex justify-center">
+      <div
+        class="flex items-center justify-center gap-2 rounded-full bg-black bg-opacity-70 px-3 py-1.5 text-sm font-bold text-white"
+      >
+        <button class="opacity-70 hover:opacity-100" @click="clickLeft">
+          <Fa6SolidAngleLeft class="text-xs" />
+        </button>
+        <p class="flex gap-1">
+          <span>{{ currentSlideIndex + 1 }}</span>
+          <span class="font-thin text-slate-300">/</span>
+          <span>{{ slides.length }}</span>
+        </p>
+        <button class="opacity-70 hover:opacity-100" @click="clickRight">
+          <Fa6SolidAngleRight class="text-xs" />
+        </button>
+      </div>
+    </div>
+    <RouterLink :to="slides[currentSlideIndex].href">
       <div
         v-for="(item, index) in slides"
         :key="index"
         class="absolute left-0 top-0 flex h-full w-full justify-center opacity-0 transition-opacity duration-1000 ease-in-out"
-        :class="
-          (index === currentSlideIndex ? 'opacity-100' : '') +
-          ' ' +
+        :class="`${index === currentSlideIndex && 'opacity-100'} ${
           bgColors[item.color]
-        "
+        }`"
       >
         <div
-          class="flex w-full max-w-7xl flex-col-reverse justify-between gap-5 p-8 md:flex-row md:px-16"
+          class="flex w-full max-w-7xl flex-col-reverse items-center justify-between gap-5 p-8 py-14 md:flex-row md:px-16 md:py-0"
         >
-          <div class="flex flex-col items-start justify-center gap-5">
+          <div class="flex flex-col items-start justify-center md:gap-5">
             <div :class="textColors[item.color]">
-              <p class="mb-2 whitespace-nowrap text-3xl font-semibold">
+              <p
+                class="mb-2 whitespace-nowrap text-2xl font-semibold md:text-3xl"
+              >
                 {{ item.topTitle }}
               </p>
-              <p class="whitespace-nowrap text-3xl font-semibold">
+              <p
+                class="mb-2 whitespace-nowrap text-2xl font-semibold md:text-3xl"
+              >
                 {{ item.bottomTitle }}
               </p>
             </div>
-            <p class="text-lg" :class="textColors[item.color]">
+            <p class="md:text-lg" :class="textColors[item.color]">
               {{ item.sub }}
             </p>
-            <div
-              class="flex items-center justify-center gap-2 rounded-full bg-black bg-opacity-70 px-3 py-1.5 text-sm font-bold text-white"
-            >
-              <button class="opacity-70 hover:opacity-100" @click="clickLeft">
-                <Fa6SolidAngleLeft class="text-xs" />
-              </button>
-              <p class="flex gap-1">
-                <span>{{ index + 1 }}</span>
-                <span class="font-thin text-slate-300">/</span>
-                <span>{{ slides.length }}</span>
-              </p>
-              <button class="opacity-70 hover:opacity-100" @click="clickRight">
-                <Fa6SolidAngleRight class="text-xs" />
-              </button>
-            </div>
           </div>
           <img
             :src="item.img"
@@ -99,6 +105,6 @@ const clickRight = () => {
           />
         </div>
       </div>
-    </div>
-  </RouterLink>
+    </RouterLink>
+  </div>
 </template>
